Move augmented beta move data into separate context
Instead of re-building the entire array of moves, we use relay everywhere in the tree (as it's intended), then the additional UI data (e.g. color and visual position) is accessible in a separate context. That way, we get the benefits of Relay everywhere.

Note: I tried using relay's client-only data functionality for this, but it just wasn't flexible enough. I ran into lots of issues with keeping the client data up-to-date with what we get from the server, especially after mutations. Turns out sticking to vanilla react features makes it a lot easier.