Remove isFree field from BetaMoveNode
This is just a derived field based on hold_id. By including it as a separate field, it makes it easy to get out-of-sync bugs in the UI. This change actually fixes one of those bugs, as the field wasn't being refreshed when mutating a move between free and unfree. By checking the presence of the related `hold` field instead, we get the same result without having to remember to refetch the derived field.