Allow annotated fields for created BetaMove rows
Previously, attemping to fetch any subquery/annotated fields (order, isStart, isLastInChain) from a beta move mutation (append OR insert) would trigger an error, because those fields are not present on the created row. We need to create the row then re-fetch from the DB with the annotations to get what we need. As part of this I also included the `order` field in the RETURNING clause. This doesn't accomplish anything *yet*, but I can see it being useful in the future.