Remove editable fields from ProblemCard
They cluttered the UI too much. These actions are available on the problem page anyway. I wanted to move the delete action to the problem page as well, but Relay makes this annoying because there's no easy way to mutate the ProblemList connection after deleting+redirecting. A problem for another day.