Fix console warning when deleting move
We were duplicating some state locally, which led to a one-frame delay after a move was deleted (or when changing betas). This led to us rendering move(s) that were no longer valid, triggering the warning. This also fixes some very minor flicker, and reduces the surface area for bugs.